Disable default output error for EXPIRATION_EARLY
This patch extends the disabling of an output error associated to
EXPIRATION_EARLY that is currently enjoyed only when a configuration
file is passed. In other terms, when no configuration file is set, any
error is returned as error code 1.

This change makes the script backwards compatible with a previous
version of the client, e.g. the client in version 0.17 had this
behaviour.