DDNS.pm: support wildcard entries
For wildcard certificates (*.$fqdn), ACME queries _acme-challenge.$fqdn,
without the "*." prefix, obviously. This has to be taken into account
when remapping the wildcard name to a different DDNS subdomain.

Moreover, when getting certificate for both $fqdn and *.$fqdn,
Crypt::LE calls ->handle_challenge_dns() twice for the same $fqdn,
and ACME expects _two_ TXT records. So we should not blindly delete
the TXT record at the beginning of ->handle_challenge_dns().

My approach is to count the number of times we've got called for
each $fqdn, store it in $self->{fqdn_seen}->{$fqdn},
and delete the TXT only in the first handle_challenge_dns() call,
and in the last handle_verification_dns() call.